Struct exoquant::optimizer::KMeans

pub struct KMeans;

A standard K-Means Optimizer.

Each palette entry is moved toward the average of the cluster of input colors it is going to represent to find a locally optimal palette.