Struct exoquant::Colorf

pub struct Colorf {
    pub r: f64,
    pub g: f64,
    pub b: f64,
    pub a: f64,
}

A color with floating point channel components.

Used for all internal processing.

It implements Mul, Div, Add and Sub operators that apply each operation component wise to each channel in turn.

Fields

r

g

b

a

Methods

impl Colorf

fn zero() -> Colorf

Returns a Colorf with all channel components set to zero.

fn dot(&self, rhs: &Colorf) -> f64

Returns the dot product of two Colorfs.

fn abs(&self) -> f64

Returns the magnitude (vector length) of a Colorf.

fn pow(&self, e: f64) -> Colorf

Returns a new Colorf with each color component raised to the given power.